requirements.txt
flask
aiogram
flask-sqlalchemy
pyjwt